refactor : api폴더 분리 및 bottomBtn common컴포넌트 제작
