+++ client/login/oauth_kakao/index.tsx
@@ -1,6 +1,6 @@
 import { useCallback, useEffect } from 'react';
 import Router from 'next/router';
-import { logInApi } from '_api';
+import { logInApi } from '_api/auth'; 
 import useLoginRedux from 'hooks/useLoginRedux';
 import useRoomRegisterRedux from 'hooks/useRoomRegisterRedux';
 import { dispatchHandleLogin } from 'store/modules/login';
